create the board resizer component

+++ Assets/Scripts/Size.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace SnakeU {
     [Serializable] public struct Size {
-        public int width;
-        public int height;
+        public float width;
+        public float height;
     }
 }
